Move each async task in the task package
